changed goal states from float[] to double[] and converted some logic to streams

+++ src/gaViz/fitness/IFitness.java
@@ -6,7 +6,7 @@ public interface IFitness {
 	
 	public void setGoal(int[] goal);
 	
-	public void setGoal(float[] goal);
+	public void setGoal(double[] goal);
 	
 	public int[] getGoal ();
